app/requirements.txt
Django==1.8.4
django-ipware==1.1.1